pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ghosttag-steganography"
version = "0.1.0"
description = "Error-Corrected Steganography for Lossy Environments"
readme = "README.md"
authors = [
  { name = "Tirthik", email = "tirthikdurgam04@gmail.com" },
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Security :: Cryptography",
    "Topic :: Multimedia :: Graphics",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "reedsolo>=1.5.0",
]

[project.urls]
"Homepage" = "https://github.com/tirthikdurgam/ghosttag"
"Bug Tracker" = "https://github.com/tirthikdurgam/ghosttag/issues"

# This section automatically finds your source code folder ("ghosttag")
[tool.setuptools.packages.find]
where = ["."]
include = ["ghosttag*"]